feat: shared/ui 타입 정의 및 파일 분리
