+++ src/shared/ui/card/CardContent.tsx
@@ -1,0 +1,10 @@
+import { forwardRef } from "react"
+
+export const CardContent = forwardRef<HTMLDivElement, React.ComponentProps<"div">>(
+  ({ className, children, ...props }, ref) => (
+    <div ref={ref} className={`p-6 pt-0 ${className}`} {...props}>
+      {children}
+    </div>
+  ),
+)
+CardContent.displayName = "CardContent"
