+++ src/shared/ui/card/CardRoot.tsx
@@ -1,0 +1,10 @@
+import { forwardRef } from "react"
+
+export const CardRoot = forwardRef<HTMLDivElement, React.ComponentProps<"div">>(
+  ({ className, children, ...props }, ref) => (
+    <div ref={ref} className={`rounded-lg border bg-card text-card-foreground shadow-sm ${className}`} {...props}>
+      {children}
+    </div>
+  ),
+)
+CardRoot.displayName = "Card"
