+++ src/shared/ui/card/CardTitle.tsx
@@ -1,0 +1,6 @@
+import { forwardRef } from "react"
+
+export const CardTitle = forwardRef<HTMLHeadingElement, React.ComponentProps<"h3">>(({ className, ...props }, ref) => (
+  <h3 ref={ref} className={`text-2xl font-semibold leading-none tracking-tight ${className}`} {...props} />
+))
+CardTitle.displayName = "CardTitle"
