+++ src/shared/ui/index.ts
@@ -1,0 +1,8 @@
+export * from "./button"
+export * from "./card"
+export * from "./dialog"
+export * from "./input"
+export * from "./table"
+export * from "./table"
+export * from "./select"
+export * from "./textarea"
