+++ src/shared/ui/select/index.ts
@@ -1,0 +1,5 @@
+export { SelectRoot as Select } from "./SelectRoot"
+export { SelectGroup, SelectValue } from "@radix-ui/react-select"
+export { SelectTrigger } from "./SelectTrigger"
+export { SelectContent } from "./SelectContent"
+export { SelectItem } from "./SelectItem"
