+++ src/shared/ui/table/TableCell.tsx
@@ -1,0 +1,8 @@
+import { forwardRef } from "react"
+
+export const TableCell = forwardRef<HTMLTableCellElement, React.ComponentProps<"td">>(
+  ({ className, ...props }, ref) => (
+    <td ref={ref} className={`p-2 align-middle [&:has([role=checkbox])]:pr-0 ${className}`} {...props} />
+  ),
+)
+TableCell.displayName = "TableCell"
