+++ src/shared/ui/table/index.ts
@@ -1,0 +1,6 @@
+export { TableRoot as Table } from "./TableRoot"
+export { TableBody } from "./TableBody"
+export { TableCell } from "./TableCell"
+export { TableHead } from "./TableHead"
+export { TableHeader } from "./TableHeader"
+export { TableRow } from "./TableRow"
